RBAC 100%

+++ saramago/backend/controllers/LeitorController.php
@@ -18,6 +18,7 @@ use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 use yii\web\Response;
 use yii\widgets\ActiveForm;
+use yii\web\ForbiddenHttpException;
 
 /**
  * LeitorController implements the CRUD actions for Leitor model.
@@ -69,44 +70,53 @@ class LeitorController extends Controller
      */
     public function actionIndex()
     {
-        $searchLeitor = new LeitorSearch();
-        $leitores = Leitor::find()->all();
-        $tiposLeitorAll = ArrayHelper::map(Tipoleitor::find()->all(),'id','designacao','tipo',['enctype' => 'multipart/form-data']);
-        $dataProvider = $searchLeitor->search(Yii::$app->request->queryParams);
-        $bibliotecasCount = Biblioteca::find()->count();
-        $tiposLeitoresCount = Tipoleitor::find()->count();
+        if ((Yii::$app->user->can('acessoLeitores'))) {
+            $searchLeitor = new LeitorSearch();
+            $leitores = Leitor::find()->all();
+            $tiposLeitorAll = ArrayHelper::map(Tipoleitor::find()->all(),'id','designacao','tipo',['enctype' => 'multipart/form-data']);
+            $dataProvider = $searchLeitor->search(Yii::$app->request->queryParams);
+            $bibliotecasCount = Biblioteca::find()->count();
+            $tiposLeitoresCount = Tipoleitor::find()->count();
 
-        $this->layout = 'minor';
+            $this->layout = 'minor';
 
-        return $this->render('index', [
-            'searchLeitor'=>$searchLeitor,
-            'leitores' => $leitores,
-            'dataProvider'=>$dataProvider,
-            'tiposLeitoresCount'=>$tiposLeitoresCount,
-            'bibliotecasCount'=>$bibliotecasCount,]);
+            return $this->render('index', [
+                'searchLeitor'=>$searchLeitor,
+                'leitores' => $leitores,
+                'dataProvider'=>$dataProvider,
+                'tiposLeitoresCount'=>$tiposLeitoresCount,
+                'bibliotecasCount'=>$bibliotecasCount,]);
+        }
+        throw new ForbiddenHttpException ('Não tem permissões para aceder à página');
     }
 
     public function actionView($id)
     {
-        $leitores = Leitor::find()->all();
+        if ((Yii::$app->user->can('acessoLeitores'))) {
+            $leitores = Leitor::find()->all();
 
-        if (($model = Leitor::findOne($id)) !== null) {
-            return $this->renderAjax('view', ['model' => $model, 'leitores'=>$leitores]);
+            if (($model = Leitor::findOne($id)) !== null) {
+                return $this->renderAjax('view', ['model' => $model, 'leitores'=>$leitores]);
+            }
+
+            return '<h5>Lamentamos! Ocorreu um erro com o seu pedido.</h5>';
         }
-
-        return '<h5>Lamentamos! Ocorreu um erro com o seu pedido.</h5>';
+        throw new ForbiddenHttpException ('Não tem permissões para aceder à página');
     }
 
     public function actionViewFull($id)
     {
-        $this->layout='minor';
-        if (($model = Leitor::findOne($id)) !== null) {
-            $user = User::findOne($model->user_id);
-            return $this->render('view-full', ['model' => $model, 'user' => $user]);
-        }
+        if ((Yii::$app->user->can('acessoLeitores'))) {
+            $this->layout='minor';
+            if (($model = Leitor::findOne($id)) !== null) {
+                $user = User::findOne($model->user_id);
+                return $this->render('view-full', ['model' => $model, 'user' => $user]);
+            }
 
-        throw new NotFoundHttpException('Leitor não encontrado.');
-    }
+            throw new NotFoundHttpException('Leitor não encontrado.');
+        }
+        throw new ForbiddenHttpException ('Não tem permissões para aceder à página');
+    }   
 
 
     /**
@@ -116,25 +126,28 @@ class LeitorController extends Controller
      */
     public function actionCreate()
     {
-        $model = new LeitorForm();
+        if ((Yii::$app->user->can('acessoLeitores'))) {
+            $model = new LeitorForm();
 
-        $listaBibliotecas = Biblioteca::find()->all();
-        $listaTiposLeitors = Tipoleitor::find()->all();
+            $listaBibliotecas = Biblioteca::find()->all();
+            $listaTiposLeitors = Tipoleitor::find()->all();
 
-        if (Yii::$app->request->isAjax && $model->load(Yii::$app->request->post()))
-        {
-            Yii::$app->response->format = Response::FORMAT_JSON;
-            return ActiveForm::validate($model);
-        }elseif ($model->load(Yii::$app->request->post()) && $model->signup()) {
-            Yii::$app->session->setFlash('success', "O Leitor foi adicionado com sucesso.");
-            return $this->redirect(['view-full', 'id' => $model->id]);
+            if (Yii::$app->request->isAjax && $model->load(Yii::$app->request->post()))
+            {
+                Yii::$app->response->format = Response::FORMAT_JSON;
+                return ActiveForm::validate($model);
+            }elseif ($model->load(Yii::$app->request->post()) && $model->signup()) {
+                Yii::$app->session->setFlash('success', "O Leitor foi adicionado com sucesso.");
+                return $this->redirect(['view-full', 'id' => $model->id]);
+            }
+
+            return $this->renderAjax('create', [
+                'model'=>$model,
+                'listaBibliotecas'=>$listaBibliotecas,
+                'listaTiposLeitors'=>$listaTiposLeitors,
+            ]);
         }
-
-        return $this->renderAjax('create', [
-            'model'=>$model,
-            'listaBibliotecas'=>$listaBibliotecas,
-            'listaTiposLeitors'=>$listaTiposLeitors,
-        ]);
+        throw new ForbiddenHttpException ('Não tem permissões para aceder à página');
     }
 
     /**
@@ -146,37 +159,43 @@ class LeitorController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = new LeitorUpdate($id);
+        if ((Yii::$app->user->can('acessoLeitores'))) {
+            $model = new LeitorUpdate($id);
 
-        $listaBibliotecas = Biblioteca::find()->all();
-        $listaTiposLeitors = Tipoleitor::find()->all();
+            $listaBibliotecas = Biblioteca::find()->all();
+            $listaTiposLeitors = Tipoleitor::find()->all();
 
-        if ($model->load(Yii::$app->request->post()) && $model->update()) {
-            Yii::$app->session->setFlash('success', "O Leitor foi editado com sucesso.");
-            return $this->redirect('..');
+            if ($model->load(Yii::$app->request->post()) && $model->update()) {
+                Yii::$app->session->setFlash('success', "O Leitor foi editado com sucesso.");
+                return $this->redirect('..');
+            }
+
+
+            return $this->renderAjax('update', [
+                'model' => $model,
+                'listaBibliotecas'=>$listaBibliotecas,
+                'listaTiposLeitors'=>$listaTiposLeitors,
+            ]);
         }
-
-
-        return $this->renderAjax('update', [
-            'model' => $model,
-            'listaBibliotecas'=>$listaBibliotecas,
-            'listaTiposLeitors'=>$listaTiposLeitors,
-        ]);
+        throw new ForbiddenHttpException ('Não tem permissões para aceder à página');
     }
 
     public function actionReporPassword($id){
-        $leitor = $this->findModel($id);
-        $user = User::findOne($leitor->user_id);
+        if ((Yii::$app->user->can('acessoLeitores'))) {
+            $leitor = $this->findModel($id);
+            $user = User::findOne($leitor->user_id);
 
-        $user->setPassword($leitor->nif);
+            $user->setPassword($leitor->nif);
 
-        if($user->save()){
-            Yii::$app->session->setFlash('success', 'A palavra passe foi reposta com sucesso');
-            return $this->redirect(['view-full', 'id' => $leitor->id]);
-        }else{
-            Yii::$app->session->setFlash('danger', 'Não foi possível repôr a palavra passe');
-            return $this->redirect(['view-full', 'id' => $leitor->id]);
+            if($user->save()){
+                Yii::$app->session->setFlash('success', 'A palavra passe foi reposta com sucesso');
+                return $this->redirect(['view-full', 'id' => $leitor->id]);
+            }else{
+                Yii::$app->session->setFlash('danger', 'Não foi possível repôr a palavra passe');
+                return $this->redirect(['view-full', 'id' => $leitor->id]);
+            }
         }
+        throw new ForbiddenHttpException ('Não tem permissões para aceder à página');
     }
 
     /**
@@ -188,15 +207,18 @@ class LeitorController extends Controller
      */
     public function actionDelete($id)
     {
-        $leitor = $this->findModel($id);
-        $user = User::findOne($leitor->user_id);
-        $user->status = 9;
+        if ((Yii::$app->user->can('acessoLeitores'))) {
+            $leitor = $this->findModel($id);
+            $user = User::findOne($leitor->user_id);
+            $user->status = 9;
 
-        $user->save();
-        $leitor->delete();
+            $user->save();
+            $leitor->delete();
 
-        Yii::$app->session->setFlash('success', "O Leitor foi eliminado com sucesso.");
-        return $this->redirect(['index']);
+            Yii::$app->session->setFlash('success', "O Leitor foi eliminado com sucesso.");
+            return $this->redirect(['index']);
+        }
+        throw new ForbiddenHttpException ('Não tem permissões para aceder à página');
     }
 
     /**
